Displays movie ID and title from database
Using PDO to pull columns from the movies database Chris made. This
displays it using recursive iterators (which looks pretty smooth in the
code!). Just messing around at this point and going to include a
dropdown form along with some php logic to filter out results in the
next update [if I can get to it tonight].